Add evidence field to ApplicationDocument and update entity path convention
- Add `evidence` array to `ApplicationDocument` schema — stores document-management
  document IDs that satisfy the checklist item, populated by the verified write-back rule
- Update `document-verified-write-back` rule to use `subjectType`/`subjectId` pattern
  (matching the appointment scheduling convention) instead of `documentId`; add
  `appendToArray` action to populate the evidence field alongside the transition
- Rename entity paths to mirror REST sub-resource hierarchy:
  `intake/application-documents` -> `intake/applications/documents`
  `intake/application-members` -> `intake/applications/members`
  `intake/interviews` -> `intake/applications/interview` (singleton, singular)
- Update entity pattern in `rules-schema.yaml` to allow multi-level paths
- Extract `deriveCollectionName` into `collection-utils.js` to break circular dependency
  (route-generator -> create-handler -> rule-evaluation -> route-generator); update all
  callers to import from the shared utility
- Regenerate intake contract tables to reflect rule and schema changes
