+++ docs/contract-tables/intake/rules-document-verified-write-back.csv
@@ -1,6 +1,8 @@
 Order,Condition,Action,Fallback,Description
-1,"{""!="":[{""var"":""document.id""},null]}","{""triggerTransition"":{""entity"":""intake/application-documents"",""idFrom"":""document.id"",""transition"":""verify""}}",,"Transition the ApplicationDocument to verified when document-management reports
-successful verification. The optional binding means rule set evaluation continues
-even if the document ID in the event does not match an intake ApplicationDocument
-(e.g., the document belongs to a different domain's checklist).
+1,"{""=="":[{""var"":""this.data.subjectType""},""application-document""]}","{""triggerTransition"":{""entity"":""intake/applications/documents"",""idFrom"":""applicationDocument.id"",""transition"":""verify""},""appendToArray"":{""entity"":""intake/applications/documents"",""idFrom"":""applicationDocument.id"",""field"":""evidence"",""value"":{""var"":""this.subject""}}}",,"Transition the ApplicationDocument to verified and record the document ID in its
+evidence list when document-management reports successful verification (Decision 15).
+Only fires when the document's subjectType is ""application-document"", ensuring
+documents uploaded in document-management without linking to an ApplicationDocument
+(via subjectType/subjectId) are silently ignored — analogous to an appointment
+scheduled without linking to an interview.
 "
